+++ src/renderer/entities/transaction/model/transaction.ts
@@ -101,11 +101,3 @@ export type MultisigTransactionKey = Pick<
   MultisigTransaction,
   'accountId' | 'callHash' | 'chainId' | 'indexCreated' | 'blockCreated'
 >;
-
-export function isDecodedTx(tx: Transaction | DecodedTransaction): tx is DecodedTransaction {
-  const hasType = tx.type !== undefined;
-  const hasMethod = (tx as DecodedTransaction).method !== undefined;
-  const hasSection = (tx as DecodedTransaction).section !== undefined;
-
-  return !hasType && hasMethod && hasSection;
-}
